- In IOFiber add a `pushCont` method, to reduce some repeated code.
- Add a method to "clear" a ByteStack, by simply changing the "size" field. Replaces the IOFiber calls to create a fiber with just "clearing" the stacks it already has.
- Some documentation on ByteStack.