@vasilmkd 2022-01-01
Can you please run `benchmarks/Jmh/run -wi 10 -i 10 -f 2 -t 1`?