@djspiewak 2022-03-18
## Before

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt      Score     Error    Units
[info] AsyncBenchmark.async                                          N/A      100  thrpt   20  32935.191 ± 388.593    ops/s
[info] AsyncBenchmark.bracket                                        N/A      100  thrpt   20  25113.693 ± 218.056    ops/s
[info] AsyncBenchmark.cancelable                                     N/A      100  thrpt   20  30370.336 ± 298.760    ops/s
[info] AsyncBenchmark.race                                           N/A      100  thrpt   20  12243.889 ± 123.191    ops/s
[info] AsyncBenchmark.racePair                                       N/A      100  thrpt   20  45134.234 ± 174.136    ops/s
[info] AsyncBenchmark.start                                          N/A      100  thrpt   20  12354.168 ± 115.447    ops/s
[info] AsyncBenchmark.uncancelable                                   N/A      100  thrpt   20  41068.163 ± 627.041    ops/s
[info] AttemptBenchmark.errorRaised                                  N/A    10000  thrpt   20   2881.860 ±   7.141    ops/s
[info] AttemptBenchmark.happyPath                                    N/A    10000  thrpt   20   5040.836 ±  10.129    ops/s
[info] BlockingBenchmark.blockThenCede                               N/A    10000  thrpt   20      7.736 ±   0.082    ops/s
[info] BlockingBenchmark.coarse                                      N/A    10000  thrpt   20   7394.527 ±   7.296    ops/s
[info] BlockingBenchmark.fine                                        N/A    10000  thrpt   20   1289.782 ±   0.999    ops/s
[info] BlockingBenchmark.nested                                      N/A    10000  thrpt   20   1261.967 ±   1.260    ops/s
[info] BothBenchmark.happyPath                                       N/A    10000  thrpt   20     50.435 ±   0.428    ops/s
[info] DeepBindBenchmark.async                                       N/A    10000  thrpt   20   2904.957 ± 134.305    ops/s
[info] DeepBindBenchmark.delay                                       N/A    10000  thrpt   20   8314.841 ±  30.544    ops/s
[info] DeepBindBenchmark.pure                                        N/A    10000  thrpt   20   9883.638 ±  27.931    ops/s
[info] DispatcherBenchmark.scheduling                                N/A     1000  thrpt   20    120.671 ±   1.139  ops/min
[info] HandleErrorBenchmark.errorRaised                              N/A    10000  thrpt   20   3441.167 ±   5.668    ops/s
[info] HandleErrorBenchmark.happyPath                                N/A    10000  thrpt   20   3970.074 ±   7.328    ops/s
[info] MapCallsBenchmark.batch120                                    N/A      N/A  thrpt   20    439.430 ±   4.297    ops/s
[info] MapCallsBenchmark.batch30                                     N/A      N/A  thrpt   20    115.857 ±   0.982    ops/s
[info] MapCallsBenchmark.one                                         N/A      N/A  thrpt   20      3.904 ±   0.028    ops/s
[info] MapStreamBenchmark.batch120                                   N/A      N/A  thrpt   20   4177.004 ±  49.343    ops/s
[info] MapStreamBenchmark.batch30                                    N/A      N/A  thrpt   20   1735.697 ±   3.748    ops/s
[info] MapStreamBenchmark.one                                        N/A      N/A  thrpt   20   2307.036 ±  82.864    ops/s
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   20    764.547 ±   1.369    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   20     68.615 ±   0.479    ops/s
[info] RaceBenchmark.happyPath                                       N/A    10000  thrpt   20     54.899 ±   1.241    ops/s
[info] RedeemBenchmark.errorRaised                                   N/A    10000  thrpt   20   2054.353 ±  46.407    ops/s
[info] RedeemBenchmark.happyPath                                     N/A    10000  thrpt   20   2654.044 ±  29.497    ops/s
[info] RedeemWithBenchmark.errorRaised                               N/A    10000  thrpt   20   3061.755 ±   6.345    ops/s
[info] RedeemWithBenchmark.happyPath                                 N/A    10000  thrpt   20   3777.555 ±  60.445    ops/s
[info] RefBenchmark.getAndUpdate                                     N/A      N/A  thrpt   20   6564.817 ±  21.048    ops/s
[info] RefBenchmark.modify                                           N/A      N/A  thrpt   20   6013.946 ±  36.104    ops/s
[info] ShallowBindBenchmark.async                                    N/A    10000  thrpt   20   2252.429 ±  12.575    ops/s
[info] ShallowBindBenchmark.delay                                    N/A    10000  thrpt   20   7172.507 ±  13.948    ops/s
[info] ShallowBindBenchmark.pure                                     N/A    10000  thrpt   20   9614.911 ±  33.521    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   20     10.691 ±   0.053  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   20     48.340 ±   2.767  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   20    170.874 ±   0.655  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   20     52.645 ±   2.552  ops/min
```

## After

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt      Score     Error    Units
[info] AsyncBenchmark.async                                          N/A      100  thrpt   20  32239.493 ± 215.580    ops/s
[info] AsyncBenchmark.bracket                                        N/A      100  thrpt   20  24552.859 ± 173.748    ops/s
[info] AsyncBenchmark.cancelable                                     N/A      100  thrpt   20  29719.106 ± 316.811    ops/s
[info] AsyncBenchmark.race                                           N/A      100  thrpt   20  12486.546 ± 374.741    ops/s
[info] AsyncBenchmark.racePair                                       N/A      100  thrpt   20  45434.388 ± 796.172    ops/s
[info] AsyncBenchmark.start                                          N/A      100  thrpt   20  12873.057 ±  79.435    ops/s
[info] AsyncBenchmark.uncancelable                                   N/A      100  thrpt   20  40112.995 ± 715.081    ops/s
[info] AttemptBenchmark.errorRaised                                  N/A    10000  thrpt   20   3324.352 ± 313.595    ops/s
[info] AttemptBenchmark.happyPath                                    N/A    10000  thrpt   20   5003.492 ±  20.690    ops/s
[info] BlockingBenchmark.blockThenCede                               N/A    10000  thrpt   20      7.657 ±   0.053    ops/s
[info] BlockingBenchmark.coarse                                      N/A    10000  thrpt   20   7308.708 ±  49.805    ops/s
[info] BlockingBenchmark.fine                                        N/A    10000  thrpt   20   1279.115 ±   3.160    ops/s
[info] BlockingBenchmark.nested                                      N/A    10000  thrpt   20   1235.176 ±   8.134    ops/s
[info] BothBenchmark.happyPath                                       N/A    10000  thrpt   20     51.706 ±   1.826    ops/s
[info] DeepBindBenchmark.async                                       N/A    10000  thrpt   20   2849.106 ± 130.486    ops/s
[info] DeepBindBenchmark.delay                                       N/A    10000  thrpt   20   8210.526 ±  43.151    ops/s
[info] DeepBindBenchmark.pure                                        N/A    10000  thrpt   20   9759.620 ±  23.823    ops/s
[info] DispatcherBenchmark.scheduling                                N/A     1000  thrpt   20    121.856 ±   0.859  ops/min
[info] HandleErrorBenchmark.errorRaised                              N/A    10000  thrpt   20   3425.628 ±  13.130    ops/s
[info] HandleErrorBenchmark.happyPath                                N/A    10000  thrpt   20   3972.367 ±  10.152    ops/s
[info] MapCallsBenchmark.batch120                                    N/A      N/A  thrpt   20    439.207 ±   2.741    ops/s
[info] MapCallsBenchmark.batch30                                     N/A      N/A  thrpt   20    113.059 ±   1.137    ops/s
[info] MapCallsBenchmark.one                                         N/A      N/A  thrpt   20      3.858 ±   0.027    ops/s
[info] MapStreamBenchmark.batch120                                   N/A      N/A  thrpt   20   4249.189 ±  17.955    ops/s
[info] MapStreamBenchmark.batch30                                    N/A      N/A  thrpt   20   1699.624 ±  23.490    ops/s
[info] MapStreamBenchmark.one                                        N/A      N/A  thrpt   20   2166.302 ±  12.895    ops/s
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   20    758.339 ±   5.141    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   20     68.181 ±   0.357    ops/s
[info] RaceBenchmark.happyPath                                       N/A    10000  thrpt   20     53.003 ±   1.104    ops/s
[info] RedeemBenchmark.errorRaised                                   N/A    10000  thrpt   20   2088.319 ±  24.363    ops/s
[info] RedeemBenchmark.happyPath                                     N/A    10000  thrpt   20   2514.424 ±   8.885    ops/s
[info] RedeemWithBenchmark.errorRaised                               N/A    10000  thrpt   20   2973.542 ±  27.798    ops/s
[info] RedeemWithBenchmark.happyPath                                 N/A    10000  thrpt   20   4427.332 ± 106.407    ops/s
[info] RefBenchmark.getAndUpdate                                     N/A      N/A  thrpt   20   6208.025 ±  32.036    ops/s
[info] RefBenchmark.modify                                           N/A      N/A  thrpt   20   5630.572 ±  24.907    ops/s
[info] ShallowBindBenchmark.async                                    N/A    10000  thrpt   20   2075.353 ±   6.998    ops/s
[info] ShallowBindBenchmark.delay                                    N/A    10000  thrpt   20   7161.117 ±  22.852    ops/s
[info] ShallowBindBenchmark.pure                                     N/A    10000  thrpt   20   9052.572 ± 459.767    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   20     10.454 ±   0.081  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   20     45.385 ±   1.596  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   20    169.730 ±   0.548  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   20     51.952 ±   2.237  ops/min
```